+++ airfoileditor/ui/xo2_dialogs.py
@@ -24,7 +24,7 @@ from model.xo2_controller   import xo2_state, Xo2_Controller
 from model.xo2_results      import Xo2_Results, Optimization_History_Entry
 from model.xo2_input        import *
 
-from ui.ae_dialogs          import Polar_Definition_Dialog
+from ui.util_dialogs        import Polar_Definition_Dialog
 from ui.ae_widgets          import Airfoil_Select_Open_Widget, mode_color
 from ui.xo2_diagrams        import Diagram_Xo2_Progress, Diagram_Xo2_Airfoil_and_Polar
